adapting imports following moving of header.jsx file
